fix: rename utils for cleanup fix: clean outdated storage 15s after map loads
